requirements_advanced.txt
bittensor
torch==2.7.1
requests
aiohttp
timelock
python-dotenv==1.1.1
boto3 
numpy
scikit-learn
scipy
matplotlib
pandas
seaborn
joblib
aiobotocore
yfinance
pandas-datareader
tqdm
xgboost
schedule

# Advanced miner dependencies
talib-binary
textblob
beautifulsoup4
feedparser
vaderSentiment
transformers
torch-audio
librosa